Add basic class structure for ETC, RBTC and CGLD
We are adding the 3 new ETH-like coins: ETC, RBTC and CGLD. This commit
adds the basic structure for these 3 coins, throwing not implemented for
most of the actual functions in the classes. This is because the logic
for those functions is still under development in the bitgo-account-lib
library and we will update here to fill in those implementations as
those are finished and published to npm. This commit helps get some of
the boiler plate out of the way when the time comes for that.

It's worth noting that ETH itself does not extend the new abstractEthLikeCoin
yet, because its implementation currently works fine. In the future when
the bitgo-account-lib implementation is finished and stable, we may
change ETH itself to use the new implementation.

Ticket: BG-17980