Update UnsupportedCoinError to to include tip
BitGoJS throws an error whenever a user tries to use a coin that is not
supported in the current version of BitGoJS. Oftentimes the fix is to
simply update BitGoJS to a version that includes the desired coin or
token. This commit adds an extra sentence to that error message to
hopefully lead users to the solution more quickly.

Ticket: BG-21130